Move package manager/state/watch into watch
