+++ manager/state/watch.go
@@ -3,7 +3,7 @@ package state
 import (
 	"github.com/docker/go-events"
 	"github.com/docker/swarmkit/api"
-	"github.com/docker/swarmkit/manager/state/watch"
+	"github.com/docker/swarmkit/watch"
 )
 
 // Event is the type used for events passed over watcher channels, and also
